Improves search bar
