+++ src/pages/layout/Search/Index.tsx
@@ -1,68 +1,229 @@
 import React, {useEffect, useState} from "react";
-import {Autocomplete, AutocompleteInputChangeReason} from "@mui/material";
+import {Autocomplete} from "@mui/material";
 import SearchInput from "./SearchInput";
-import useGetSearchResults, {
-  NotFoundResult,
-  SearchResult,
-} from "../../../api/hooks/useGetSearchResults";
 import ResultLink from "./ResultLink";
 import {
   useAugmentToWithGlobalSearchParams,
   useNavigate,
 } from "../../../routing";
+import {useGlobalState} from "../../../global-config/GlobalConfig";
+import {GTMEvents} from "../../../dataConstants";
+import {
+  getAccount,
+  getAccountResources,
+  getTransaction,
+  getBlockByHeight,
+  getBlockByVersion,
+} from "../../../api";
+import {sendToGTM} from "../../../api/hooks/useGoogleTagManager";
+import {objectCoreAddress} from "../../../constants";
+import {
+  isValidAccountAddress,
+  isValidTxnHashOrVersion,
+  isNumeric,
+  truncateAddress,
+} from "../../utils";
+
+export type SearchResult = {
+  label: string;
+  to: string | null;
+};
+
+export const NotFoundResult: SearchResult = {
+  label: "No Results",
+  to: null,
+};
+
+type SearchMode = "idle" | "typing" | "loading" | "results";
 
 export default function HeaderSearch() {
   const navigate = useNavigate();
+  const [state] = useGlobalState();
+  const [mode, setMode] = useState<SearchMode>("idle");
   const [inputValue, setInputValue] = useState<string>("");
-  const [searchValue, setSearchValue] = useState<string>("");
+  const [options, setOptions] = useState<SearchResult[]>([]);
   const [open, setOpen] = useState(false);
-  const [selectedOption, setSelectedOption] =
-    useState<SearchResult>(NotFoundResult);
+  const [selectedOption, setSelectedOption] = useState<SearchResult | null>(
+    null,
+  );
   const augmentToWithGlobalSearchParams = useAugmentToWithGlobalSearchParams();
 
-  const options = useGetSearchResults(searchValue).map((result) => ({
-    ...result,
-    to: result.to !== null ? augmentToWithGlobalSearchParams(result.to) : null,
-  }));
-
-  // inputValue is the value in the text field
-  // searchValue is the value that we search
-  // searchValue is updated 0.5s after the inputValue is changed
-  // this is to wait for users to stop typing then execute searching
   useEffect(() => {
-    const timer = setTimeout(() => {
-      setSearchValue(inputValue);
-    }, 500);
+    let timer: NodeJS.Timeout;
+
+    if (mode !== "loading" && inputValue.trim().length > 0) {
+      timer = setTimeout(() => {
+        fetchData(inputValue.trim());
+      }, 500);
+    }
 
     return () => clearTimeout(timer);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [inputValue]);
 
-  const handleInputChange = (
-    event: any,
-    newInputValue: React.SetStateAction<string>,
-    reason: AutocompleteInputChangeReason,
-  ) => {
-    if (newInputValue.length === 0) {
-      if (open) {
-        setOpen(false);
-      }
+  const fetchData = async (searchText: string) => {
+    setMode("loading");
+    const searchPerformanceStart = GTMEvents.SEARCH_STATS + " start";
+    const searchPerformanceEnd = GTMEvents.SEARCH_STATS + " end";
+    window.performance.mark(searchPerformanceStart);
+
+    const isValidAccountAddr = isValidAccountAddress(searchText);
+    const isValidTxnHashOrVer = isValidTxnHashOrVersion(searchText);
+    const isValidBlockHeightOrVer = isNumeric(searchText);
+
+    const promises = [];
+
+    if (searchText.endsWith(".apt")) {
+      try {
+        const name = await state.sdk_v2_client?.getName({
+          name: searchText,
+        });
+        const address = name?.registered_address;
+        const primaryName = await state.sdk_v2_client?.getPrimaryName({
+          address: name?.owner_address ?? "",
+        });
+
+        if (!primaryName || !name || !address) {
+          throw new Error("Primary name not found");
+        }
+
+        promises.push({
+          label: `Account ${truncateAddress(address)}${
+            primaryName ? ` | ${primaryName}.apt` : ``
+          }`,
+          to: `/account/${name.owner_address}`,
+        });
+      } catch (e) {}
     } else {
-      if (!open) {
-        setOpen(true);
+      if (isValidAccountAddr) {
+        // It's either an account OR an object: we query both at once to save time
+        const accountPromise = await getAccount(
+          {address: searchText},
+          state.network_value,
+        )
+          .then((): SearchResult => {
+            return {
+              label: `Account ${searchText}`,
+              to: `/account/${searchText}`,
+            };
+          })
+          .catch(() => {
+            return null;
+            // Do nothing. It's expected that not all search input is a valid account
+          });
+
+        const resourcePromise = await getAccountResources(
+          {address: searchText},
+          state.network_value,
+        )
+          .then((resources): SearchResult | undefined => {
+            let hasObjectCore = false;
+            resources.forEach((resource) => {
+              if (resource.type === objectCoreAddress) {
+                hasObjectCore = true;
+              }
+            });
+            if (hasObjectCore) {
+              return {
+                label: `Object ${searchText}`,
+                to: `/object/${searchText}`,
+              };
+            }
+          })
+          .catch(() => {
+            return null;
+            // Do nothing. It's expected that not all search input is a valid account
+          });
+        promises.push(accountPromise);
+        promises.push(resourcePromise);
+      }
+
+      if (isValidTxnHashOrVer) {
+        const txnPromise = getTransaction(
+          {txnHashOrVersion: searchText},
+          state.network_value,
+        )
+          .then((): SearchResult => {
+            return {
+              label: `Transaction ${searchText}`,
+              to: `/txn/${searchText}`,
+            };
+          })
+          .catch(() => {
+            return null;
+            // Do nothing. It's expected that not all search input is a valid transaction
+          });
+        promises.push(txnPromise);
+      }
+
+      if (isValidBlockHeightOrVer) {
+        const blockByHeightPromise = getBlockByHeight(
+          {height: parseInt(searchText), withTransactions: false},
+          state.network_value,
+        )
+          .then((): SearchResult => {
+            return {
+              label: `Block ${searchText}`,
+              to: `/block/${searchText}`,
+            };
+          })
+          .catch(() => {
+            return null;
+            // Do nothing. It's expected that not all search input is a valid transaction
+          });
+
+        const blockByVersionPromise = getBlockByVersion(
+          {version: parseInt(searchText), withTransactions: false},
+          state.network_value,
+        )
+          .then((block): SearchResult => {
+            return {
+              label: `Block with Txn Version ${searchText}`,
+              to: `/block/${block.block_height}`,
+            };
+          })
+          .catch(() => {
+            return null;
+            // Do nothing. It's expected that not all search input is a valid transaction
+          });
+        promises.push(blockByHeightPromise);
+        promises.push(blockByVersionPromise);
       }
     }
 
-    if (event && event.type === "blur") {
-      setInputValue("");
-    } else if (reason !== "reset") {
-      setInputValue(newInputValue);
-    }
-  };
+    const resultsList = await Promise.all(promises);
+    const results = resultsList
+      .filter((result): result is SearchResult => !!result)
+      .map((result) => {
+        if (result.to) {
+          return {...result, to: augmentToWithGlobalSearchParams(result.to)};
+        }
 
-  const handleSubmitSearch = async () => {
-    if (selectedOption.to !== null) {
-      navigate(selectedOption.to);
+        return result;
+      });
+
+    window.performance.mark(searchPerformanceEnd);
+    sendToGTM({
+      dataLayer: {
+        event: GTMEvents.SEARCH_STATS,
+        network: state.network_name,
+        searchText: searchText,
+        searchResult: results.length === 0 ? "notFound" : "success",
+        duration: window.performance.measure(
+          GTMEvents.SEARCH_STATS,
+          searchPerformanceStart,
+          searchPerformanceEnd,
+        ).duration,
+      },
+    });
+
+    if (results.length === 0) {
+      results.push(NotFoundResult);
     }
+
+    setOptions(results);
+    setMode("idle");
+    setOpen(true);
   };
 
   return (
@@ -98,10 +259,32 @@ export default function HeaderSearch() {
       filterOptions={(x) => x.filter((x) => !!x)}
       options={options}
       inputValue={inputValue}
-      onInputChange={handleInputChange}
+      onInputChange={(event, newInputValue, reason) => {
+        setOpen(false);
+        if (event && event.type === "blur") {
+          setInputValue("");
+        } else if (reason !== "reset") {
+          setMode(newInputValue.trim().length === 0 ? "idle" : "typing");
+          setInputValue(newInputValue);
+        }
+      }}
+      onKeyDown={(event) => {
+        if (event.key === "Enter") {
+          const selected = selectedOption?.to ?? options[0]?.to;
+          if (selected) {
+            navigate(selected);
+          }
+          event.preventDefault();
+        }
+      }}
       onClose={() => setOpen(false)}
       renderInput={(params) => {
-        return <SearchInput {...params} />;
+        return (
+          <SearchInput
+            {...params}
+            loading={mode === "loading" || mode === "typing"}
+          />
+        );
       }}
       renderOption={(props, option) => {
         return (
@@ -114,12 +297,6 @@ export default function HeaderSearch() {
         if (option !== null) {
           const optionCopy = Object.assign({}, option);
           setSelectedOption(optionCopy);
-        }
-      }}
-      onKeyDown={(event) => {
-        if (event.key === "Enter") {
-          handleSubmitSearch();
-          event.preventDefault();
         }
       }}
     />
