+++ src/pages/layout/Search/SearchInput.tsx
@@ -1,8 +1,17 @@
 import React from "react";
-import {InputAdornment, TextField} from "@mui/material";
+import {
+  AutocompleteRenderInputParams,
+  CircularProgress,
+  InputAdornment,
+  TextField,
+} from "@mui/material";
 import SearchIcon from "@mui/icons-material/Search";
 
-export default function SearchInput({...params}) {
+interface SearchInputProps extends AutocompleteRenderInputParams {
+  loading?: boolean;
+}
+
+export default function SearchInput({loading, ...params}: SearchInputProps) {
   return (
     <form style={{width: "100%"}}>
       <TextField
@@ -20,6 +29,11 @@ export default function SearchInput({...params}) {
               sx={{ml: 0.5, marginTop: "0!important"}}
             >
               <SearchIcon fontSize="large" color="secondary" />
+            </InputAdornment>
+          ),
+          endAdornment: loading && (
+            <InputAdornment position="end">
+              <CircularProgress size={20} />
             </InputAdornment>
           ),
         }}
